pyproject.toml
[tool.poetry]
name = "lifx_cli"
version = "0.1.0"
description = "CLI App for interacting with Lifx bulbs"
authors = ["Harry Cooper <harry.coop99@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
click = "^7.1"
lifx-wrapper = "^0.1.0"
typer = "^0.2.1"
colorama = "^0.4.3"
configobj = "^5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
